+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.11.0] – 2021-03-05
+
+### New
+- [`utils.calc_storage_fee`](docs/mod_utils.md#calc_storage_fee) function to calculate account storage fee over a some time period.
+- **Debot Module**:
+    - Added unstable functions to `Sdk` interface: `getAccountsDataByHash`
+
 ## [1.10.0] – 2021-03-04
 
 ### New